Moved email check from adapter to a simple util method. Altered signup to check for email being a valid type. Updated import method
